refactor : 불필요한 컴포넌트 제거 , 폴더 정리
폴더 및 컴포넌트 정리 했습니다.
