--- openvidu-call-react/src/SocketIo.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-const SocketIo = () => {
-  return (
-    <div>
-      <h1>소켓서버</h1>
-    </div>
-  );
-};
-
-export default SocketIo;
