+++ openvidu-call-react/src/components/recognition/Dictaphone.js
