+++ openvidu-call-react/src/components/recognition/Recognition.js
