+++ openvidu-call-react/src/components/videoroom/VideoRoomComponent.css
