+++ openvidu-call-react/src/index.js
@@ -1,24 +1,21 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import "./index.css";
-import VideoRoomComponent from "./components/VideoRoomComponent";
+import VideoRoomHandsFree from "./components/videoroom/VideoRoomHandsFree";
 import registerServiceWorker from "./registerServiceWorker";
 import App from "./App";
 import Main from "./main/main";
 import { BrowserRouter } from "react-router-dom";
 import { Provider } from "react-redux";
-import store from './store.js'
-
+import store from "./store.js";
 
 ReactDOM.render(
   <Provider store={store}>
     <BrowserRouter>
       <App />
     </BrowserRouter>
-  </Provider>
-  ,
+  </Provider>,
   document.getElementById("root")
 );
-
 
 registerServiceWorker();
